Restructure Code
