+++ src/PhpSpreadsheet/Reader/Xlsx.php
@@ -53,24 +53,16 @@ class Xlsx extends BaseReader
     /**
      * ReferenceHelper instance.
      */
-    private ReferenceHelper $referenceHelper;
+    protected ReferenceHelper $referenceHelper;
 
-    private ZipArchive $zip;
+    protected ZipArchive $zip;
 
     private Styles $styleReader;
 
     /** @var SharedFormula[] */
-    private array $sharedFormulae = [];
+    protected array $sharedFormulae = [];
 
-    private bool $parseHuge = false;
-
-    /**
-     * Use XMLReader-based streaming for cell data parsing.
-     * When enabled, worksheet cell data is parsed using XMLReader
-     * instead of SimpleXML, which significantly reduces memory usage
-     * for large worksheets.
-     */
-    private bool $useStreamingReader = false;
+    protected bool $parseHuge = false;
 
     /**
      * Allow use of LIBXML_PARSEHUGE.
@@ -81,18 +73,6 @@ class Xlsx extends BaseReader
     public function setParseHuge(bool $parseHuge): void
     {
         $this->parseHuge = $parseHuge;
-    }
-
-    public function getUseStreamingReader(): bool
-    {
-        return $this->useStreamingReader;
-    }
-
-    public function setUseStreamingReader(bool $useStreamingReader): self
-    {
-        $this->useStreamingReader = $useStreamingReader;
-
-        return $this;
     }
 
     /**
@@ -339,7 +319,7 @@ class Xlsx extends BaseReader
         return $worksheetInfo;
     }
 
-    private static function castToBoolean(SimpleXMLElement $c): bool
+    protected static function castToBoolean(SimpleXMLElement $c): bool
     {
         $value = isset($c->v) ? (string) $c->v : null;
         if ($value == '0') {
@@ -351,12 +331,12 @@ class Xlsx extends BaseReader
         return (bool) $c->v;
     }
 
-    private static function castToError(?SimpleXMLElement $c): ?string
+    protected static function castToError(?SimpleXMLElement $c): ?string
     {
         return isset($c, $c->v) ? (string) $c->v : null;
     }
 
-    private static function castToString(?SimpleXMLElement $c): ?string
+    protected static function castToString(?SimpleXMLElement $c): ?string
     {
         return isset($c, $c->v) ? (string) $c->v : null;
     }
@@ -366,7 +346,7 @@ class Xlsx extends BaseReader
         return str_replace(['_xlfn.', '_xlws.'], '', $formula);
     }
 
-    private function castToFormula(?SimpleXMLElement $c, string $r, string &$cellDataType, mixed &$value, mixed &$calculatedValue, string $castBaseType, bool $updateSharedCells = true): void
+    protected function castToFormula(?SimpleXMLElement $c, string $r, string &$cellDataType, mixed &$value, mixed &$calculatedValue, string $castBaseType, bool $updateSharedCells = true): void
     {
         if ($c === null) {
             return;
@@ -418,7 +398,7 @@ class Xlsx extends BaseReader
         return $contents !== false;
     }
 
-    private function getFromZipArchive(ZipArchive $archive, string $fileName = ''): string
+    protected function getFromZipArchive(ZipArchive $archive, string $fileName = ''): string
     {
         // Root-relative paths
         if (str_contains($fileName, '//')) {
@@ -914,193 +894,19 @@ class Xlsx extends BaseReader
                             }
 
                             $holdSelectedCells = $docSheet->getSelectedCells();
-                            if ($this->useStreamingReader) {
-                                /** @var array<object> $styles */
-                                $this->loadSheetDataWithXmlReader(
-                                    $docSheet,
-                                    "$dir/$fileWorksheet",
-                                    $mainNS,
-                                    $sharedStrings,
-                                    $styles
-                                );
-                            } elseif ($xmlSheetNS && $xmlSheetNS->sheetData && $xmlSheetNS->sheetData->row) {
-                                $cIndex = 1; // Cell Start from 1
-                                foreach ($xmlSheetNS->sheetData->row as $row) {
-                                    $rowIndex = 1;
-                                    foreach ($row->c as $c) {
-                                        $cAttr = self::getAttributes($c);
-                                        $r = (string) $cAttr['r'];
-                                        if ($r == '') {
-                                            $r = Coordinate::stringFromColumnIndex($rowIndex) . $cIndex;
-                                        }
-                                        $cellDataType = (string) $cAttr['t'];
-                                        $originalCellDataTypeNumeric = $cellDataType === '';
-                                        $value = null;
-                                        $calculatedValue = null;
+                            /** @var array<object> $styles */
+                            $this->loadSheetData(
+                                $xmlSheetNS,
+                                $filename,
+                                $dir,
+                                $richData,
+                                $docSheet,
+                                "$dir/$fileWorksheet",
+                                $mainNS,
+                                $sharedStrings,
+                                $styles
+                            );
 
-                                        // Read cell?
-                                        $coordinates = Coordinate::coordinateFromString($r);
-
-                                        if (!$this->getReadFilter()->readCell($coordinates[0], (int) $coordinates[1], $docSheet->getTitle())) {
-                                            // Normally, just testing for the f attribute should identify this cell as containing a formula
-                                            // that we need to read, even though it is outside of the filter range, in case it is a shared formula.
-                                            // But in some cases, this attribute isn't set; so we need to delve a level deeper and look at
-                                            // whether or not the cell has a child formula element that is shared.
-                                            if (isset($cAttr->f) || (isset($c->f, $c->f->attributes()['t']) && strtolower((string) $c->f->attributes()['t']) === 'shared')) {
-                                                $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToError', false);
-                                            }
-                                            ++$rowIndex;
-
-                                            continue;
-                                        }
-
-                                        // Read cell!
-                                        $useFormula = isset($c->f)
-                                            && ((string) $c->f !== '' || (isset($c->f->attributes()['t']) && strtolower((string) $c->f->attributes()['t']) === 'shared'));
-                                        switch ($cellDataType) {
-                                            case DataType::TYPE_STRING:
-                                                if ((string) $c->v != '') {
-                                                    $value = $sharedStrings[(int) ($c->v)];
-
-                                                    if ($value instanceof RichText) {
-                                                        $value = clone $value;
-                                                    }
-                                                } else {
-                                                    $value = '';
-                                                }
-
-                                                break;
-                                            case DataType::TYPE_BOOL:
-                                                if (!$useFormula) {
-                                                    if (isset($c->v)) {
-                                                        $value = self::castToBoolean($c);
-                                                    } else {
-                                                        $value = null;
-                                                        $cellDataType = DataType::TYPE_NULL;
-                                                    }
-                                                } else {
-                                                    // Formula
-                                                    $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToBoolean');
-                                                    self::storeFormulaAttributes($c->f, $docSheet, $r);
-                                                }
-
-                                                break;
-                                            case DataType::TYPE_STRING2:
-                                                if ($useFormula) {
-                                                    $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToString');
-                                                    self::storeFormulaAttributes($c->f, $docSheet, $r);
-                                                } else {
-                                                    $value = self::castToString($c);
-                                                }
-
-                                                break;
-                                            case DataType::TYPE_INLINE:
-                                                if ($useFormula) {
-                                                    $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToError');
-                                                    self::storeFormulaAttributes($c->f, $docSheet, $r);
-                                                } else {
-                                                    $value = $this->parseRichText($c->is);
-                                                }
-
-                                                break;
-                                            case DataType::TYPE_ERROR:
-                                                if (isset($cAttr->vm, $richData['image']['rId' . $cAttr->vm]) && !$useFormula) {
-                                                    $imagePath = $dir . '/' . str_replace('../', '', $richData['image']['rId' . $cAttr->vm]);
-                                                    $objDrawing = new \PhpOffice\PhpSpreadsheet\Worksheet\Drawing();
-                                                    $objDrawing->setPath(
-                                                        'zip://' . File::realpath($filename) . '#' . $imagePath,
-                                                        false,
-                                                        $zip
-                                                    );
-
-                                                    $objDrawing->setCoordinates($r);
-                                                    $objDrawing->setResizeProportional(false);
-                                                    $objDrawing->setInCell(true);
-                                                    $objDrawing->setWorksheet($docSheet);
-
-                                                    $value = $objDrawing;
-                                                    $cellDataType = DataType::TYPE_DRAWING_IN_CELL;
-                                                    $c->t = DataType::TYPE_ERROR;
-
-                                                    break;
-                                                }
-
-                                                if (!$useFormula) {
-                                                    $value = self::castToError($c);
-                                                } else {
-                                                    // Formula
-                                                    $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToError');
-                                                    $eattr = $c->attributes();
-                                                    if (isset($eattr['vm'])) {
-                                                        if ($calculatedValue === ExcelError::VALUE()) {
-                                                            $calculatedValue = ExcelError::SPILL();
-                                                        }
-                                                    }
-                                                }
-
-                                                break;
-                                            default:
-                                                if (!$useFormula) {
-                                                    $value = self::castToString($c);
-                                                    if (is_numeric($value)) {
-                                                        $value += 0;
-                                                        $cellDataType = DataType::TYPE_NUMERIC;
-                                                    }
-                                                } else {
-                                                    // Formula
-                                                    $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToString');
-                                                    if (is_numeric($calculatedValue)) {
-                                                        $calculatedValue += 0;
-                                                    }
-                                                    self::storeFormulaAttributes($c->f, $docSheet, $r);
-                                                }
-
-                                                break;
-                                        }
-
-                                        // read empty cells or the cells are not empty
-                                        if ($this->readEmptyCells || ($value !== null && $value !== '')) {
-                                            // Rich text?
-                                            if ($value instanceof RichText && $this->readDataOnly) {
-                                                $value = $value->getPlainText();
-                                            }
-
-                                            $cell = $docSheet->getCell($r);
-                                            // Assign value
-                                            if ($cellDataType != '') {
-                                                // it is possible, that datatype is numeric but with an empty string, which result in an error
-                                                if ($cellDataType === DataType::TYPE_NUMERIC && ($value === '' || $value === null)) {
-                                                    $cellDataType = DataType::TYPE_NULL;
-                                                }
-                                                if ($cellDataType !== DataType::TYPE_NULL) {
-                                                    $cell->setValueExplicit($value, $cellDataType);
-                                                }
-                                            } else {
-                                                $cell->setValue($value);
-                                            }
-                                            if ($calculatedValue !== null) {
-                                                $cell->setCalculatedValue($calculatedValue, $originalCellDataTypeNumeric);
-                                            }
-
-                                            // Style information?
-                                            if (!$this->readDataOnly) {
-                                                $cAttrS = (int) ($cAttr['s'] ?? 0);
-                                                // no style index means 0, it seems
-                                                $cAttrS = isset($styles[$cAttrS]) ? $cAttrS : 0;
-                                                $cell->setXfIndex($cAttrS);
-                                                // issue 3495
-                                                if ($cellDataType === DataType::TYPE_FORMULA && $styles[$cAttrS]->quotePrefix === true) { //* @phpstan-ignore-line
-                                                    $holdSelected = $docSheet->getSelectedCells();
-                                                    $cell->getStyle()->setQuotePrefix(false);
-                                                    $docSheet->setSelectedCells($holdSelected);
-                                                }
-                                            }
-                                        }
-                                        ++$rowIndex;
-                                    }
-                                    ++$cIndex;
-                                }
-                            }
                             $docSheet->setSelectedCells($holdSelectedCells);
                             if (!$this->readDataOnly && $xmlSheetNS && $xmlSheetNS->ignoredErrors) {
                                 foreach ($xmlSheetNS->ignoredErrors->ignoredError as $ignoredError) {
@@ -2084,7 +1890,208 @@ class Xlsx extends BaseReader
         return $excel;
     }
 
-    private function parseRichText(?SimpleXMLElement $is): RichText
+    /**
+     * @param string[][] $richData
+     * @param Worksheet $docSheet the worksheet to populate
+     * @param string $fileWorksheetPath path to worksheet XML within the zip
+     * @param string $mainNS the main spreadsheetml namespace
+     * @param array<int, mixed> $sharedStrings shared string table
+     * @param object[] $styles style objects array
+     */
+    protected function loadSheetData(
+        ?SimpleXMLElement $xmlSheetNS,
+        string $filename,
+        string $dir,
+        array $richData,
+        Worksheet $docSheet,
+        string $fileWorksheetPath,
+        string $mainNS,
+        array $sharedStrings,
+        array $styles,
+    ): void {
+        if (!($xmlSheetNS && $xmlSheetNS->sheetData && $xmlSheetNS->sheetData->row)) {
+            return; // @codeCoverageIgnore
+        }
+
+        $cIndex = 1; // Cell Start from 1
+        foreach ($xmlSheetNS->sheetData->row as $row) {
+            $rowIndex = 1;
+            foreach ($row->c as $c) {
+                $cAttr = self::getAttributes($c);
+                $r = (string) $cAttr['r'];
+                if ($r == '') {
+                    $r = Coordinate::stringFromColumnIndex($rowIndex) . $cIndex;
+                }
+                $cellDataType = (string) $cAttr['t'];
+                $originalCellDataTypeNumeric = $cellDataType === '';
+                $value = null;
+                $calculatedValue = null;
+
+                // Read cell?
+                $coordinates = Coordinate::coordinateFromString($r);
+
+                if (!$this->getReadFilter()->readCell($coordinates[0], (int) $coordinates[1], $docSheet->getTitle())) {
+                    // Normally, just testing for the f attribute should identify this cell as containing a formula
+                    // that we need to read, even though it is outside of the filter range, in case it is a shared formula.
+                    // But in some cases, this attribute isn't set; so we need to delve a level deeper and look at
+                    // whether or not the cell has a child formula element that is shared.
+                    if (isset($cAttr->f) || (isset($c->f, $c->f->attributes()['t']) && strtolower((string) $c->f->attributes()['t']) === 'shared')) {
+                        $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToError', false);
+                    }
+                    ++$rowIndex;
+
+                    continue;
+                }
+
+                // Read cell!
+                $useFormula = isset($c->f)
+                    && ((string) $c->f !== '' || (isset($c->f->attributes()['t']) && strtolower((string) $c->f->attributes()['t']) === 'shared'));
+                switch ($cellDataType) {
+                    case DataType::TYPE_STRING:
+                        if ((string) $c->v != '') {
+                            $value = $sharedStrings[(int) ($c->v)];
+
+                            if ($value instanceof RichText) {
+                                $value = clone $value;
+                            }
+                        } else {
+                            $value = '';
+                        }
+
+                        break;
+                    case DataType::TYPE_BOOL:
+                        if (!$useFormula) {
+                            if (isset($c->v)) {
+                                $value = self::castToBoolean($c);
+                            } else {
+                                $value = null;
+                                $cellDataType = DataType::TYPE_NULL;
+                            }
+                        } else {
+                            // Formula
+                            $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToBoolean');
+                            self::storeFormulaAttributes($c->f, $docSheet, $r);
+                        }
+
+                        break;
+                    case DataType::TYPE_STRING2:
+                        if ($useFormula) {
+                            $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToString');
+                            self::storeFormulaAttributes($c->f, $docSheet, $r);
+                        } else {
+                            $value = self::castToString($c);
+                        }
+
+                        break;
+                    case DataType::TYPE_INLINE:
+                        if ($useFormula) {
+                            $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToError');
+                            self::storeFormulaAttributes($c->f, $docSheet, $r);
+                        } else {
+                            $value = $this->parseRichText($c->is);
+                        }
+
+                        break;
+                    case DataType::TYPE_ERROR:
+                        if (isset($cAttr->vm, $richData['image']['rId' . $cAttr->vm]) && !$useFormula) {
+                            $imagePath = $dir . '/' . str_replace('../', '', $richData['image']['rId' . $cAttr->vm]);
+                            $objDrawing = new \PhpOffice\PhpSpreadsheet\Worksheet\Drawing();
+                            $objDrawing->setPath(
+                                'zip://' . File::realpath($filename) . '#' . $imagePath,
+                                false,
+                                $this->zip
+                            );
+
+                            $objDrawing->setCoordinates($r);
+                            $objDrawing->setResizeProportional(false);
+                            $objDrawing->setInCell(true);
+                            $objDrawing->setWorksheet($docSheet);
+
+                            $value = $objDrawing;
+                            $cellDataType = DataType::TYPE_DRAWING_IN_CELL;
+                            $c->t = DataType::TYPE_ERROR;
+
+                            break;
+                        }
+
+                        if (!$useFormula) {
+                            $value = self::castToError($c);
+                        } else {
+                            // Formula
+                            $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToError');
+                            $eattr = $c->attributes();
+                            if (isset($eattr['vm'])) {
+                                if ($calculatedValue === ExcelError::VALUE()) {
+                                    $calculatedValue = ExcelError::SPILL();
+                                }
+                            }
+                        }
+
+                        break;
+                    default:
+                        if (!$useFormula) {
+                            $value = self::castToString($c);
+                            if (is_numeric($value)) {
+                                $value += 0;
+                                $cellDataType = DataType::TYPE_NUMERIC;
+                            }
+                        } else {
+                            // Formula
+                            $this->castToFormula($c, $r, $cellDataType, $value, $calculatedValue, 'castToString');
+                            if (is_numeric($calculatedValue)) {
+                                $calculatedValue += 0;
+                            }
+                            self::storeFormulaAttributes($c->f, $docSheet, $r);
+                        }
+
+                        break;
+                }
+
+                // read empty cells or the cells are not empty
+                if ($this->readEmptyCells || ($value !== null && $value !== '')) {
+                    // Rich text?
+                    if ($value instanceof RichText && $this->readDataOnly) {
+                        $value = $value->getPlainText();
+                    }
+
+                    $cell = $docSheet->getCell($r);
+                    // Assign value
+                    if ($cellDataType != '') {
+                        // it is possible, that datatype is numeric but with an empty string, which result in an error
+                        if ($cellDataType === DataType::TYPE_NUMERIC && ($value === '' || $value === null)) {
+                            $cellDataType = DataType::TYPE_NULL;
+                        }
+                        if ($cellDataType !== DataType::TYPE_NULL) {
+                            $cell->setValueExplicit($value, $cellDataType);
+                        }
+                    } else {
+                        $cell->setValue($value);
+                    }
+                    if ($calculatedValue !== null) {
+                        $cell->setCalculatedValue($calculatedValue, $originalCellDataTypeNumeric);
+                    }
+
+                    // Style information?
+                    if (!$this->readDataOnly) {
+                        $cAttrS = (int) ($cAttr['s'] ?? 0);
+                        // no style index means 0, it seems
+                        $cAttrS = isset($styles[$cAttrS]) ? $cAttrS : 0;
+                        $cell->setXfIndex($cAttrS);
+                        // issue 3495
+                        if ($cellDataType === DataType::TYPE_FORMULA && $styles[$cAttrS]->quotePrefix === true) { //* @phpstan-ignore-line
+                            $holdSelected = $docSheet->getSelectedCells();
+                            $cell->getStyle()->setQuotePrefix(false);
+                            $docSheet->setSelectedCells($holdSelected);
+                        }
+                    }
+                }
+                ++$rowIndex;
+            }
+            ++$cIndex;
+        }
+    }
+
+    protected function parseRichText(?SimpleXMLElement $is): RichText
     {
         $value = new RichText();
 
@@ -2695,346 +2702,7 @@ class Xlsx extends BaseReader
         }
     }
 
-    /**
-     * Load sheet data using XMLReader for memory-efficient streaming.
-     *
-     * @param Worksheet $docSheet the worksheet to populate
-     * @param string $fileWorksheetPath path to worksheet XML within the zip
-     * @param string $mainNS the main spreadsheetml namespace
-     * @param array<int, mixed> $sharedStrings shared string table
-     * @param object[] $styles style objects array
-     */
-    private function loadSheetDataWithXmlReader(
-        Worksheet $docSheet,
-        string $fileWorksheetPath,
-        string $mainNS,
-        array $sharedStrings,
-        array $styles,
-    ): void {
-        $xmlContent = $this->getFromZipArchive($this->zip, $fileWorksheetPath);
-        if ($xmlContent === '') {
-            return; // @codeCoverageIgnore
-        }
-
-        $xml = new XMLReader();
-        $xml->xml(
-            $this->getSecurityScannerOrThrow()->scan($xmlContent),
-            null,
-            $this->parseHuge ? LIBXML_PARSEHUGE : 0
-        );
-        $xml->setParserProperty(XMLReader::SUBST_ENTITIES, true);
-
-        // Free the raw XML string now that XMLReader has consumed it
-        unset($xmlContent);
-
-        $inSheetData = false;
-        $cIndex = 1;
-
-        while ($xml->read()) {
-            // Look for the sheetData element to start processing
-            if ($xml->localName === 'sheetData' && $xml->namespaceURI === $mainNS) {
-                if ($xml->nodeType === XMLReader::ELEMENT) {
-                    $inSheetData = true;
-                    if ($xml->isEmptyElement) {
-                        break; // empty sheetData @codeCoverageIgnore
-                    }
-                } elseif ($xml->nodeType === XMLReader::END_ELEMENT) {
-                    break; // end of sheetData
-                }
-
-                continue;
-            }
-
-            if (!$inSheetData) {
-                continue;
-            }
-
-            // Process row elements
-            if ($xml->localName === 'row' && $xml->nodeType === XMLReader::ELEMENT && $xml->namespaceURI === $mainNS) {
-                $rowIndex = 1;
-
-                if ($xml->isEmptyElement) { // @codeCoverageIgnoreStart
-                    ++$cIndex;
-
-                    continue; // @codeCoverageIgnoreEnd
-                }
-
-                // Read cell elements within this row
-                while ($xml->read()) {
-                    if ($xml->localName === 'row' && $xml->nodeType === XMLReader::END_ELEMENT) {
-                        break; // end of row
-                    }
-
-                    if ($xml->localName === 'c' && $xml->nodeType === XMLReader::ELEMENT && $xml->namespaceURI === $mainNS) {
-                        $r = $xml->getAttribute('r') ?? '';
-                        if ($r === '') {
-                            $r = Coordinate::stringFromColumnIndex($rowIndex) . $cIndex; // @codeCoverageIgnore
-                        }
-                        $cellDataType = $xml->getAttribute('t') ?? '';
-                        $originalCellDataTypeNumeric = $cellDataType === '';
-                        $styleIndex = (int) ($xml->getAttribute('s') ?? 0);
-                        $value = null;
-                        $calculatedValue = null;
-
-                        // Read cell?
-                        $coordinates = Coordinate::coordinateFromString($r);
-
-                        // Parse the cell's inner XML using SimpleXML for the <c> element subtree only.
-                        // This keeps memory-efficient row-by-row iteration while reusing
-                        // existing parsing logic for formulas, inline strings, etc.
-                        $cellXml = null;
-                        if (!$xml->isEmptyElement) {
-                            $outerXml = $xml->readOuterXml();
-                            if ($outerXml !== '') {
-                                // readOuterXml() typically includes inherited namespace declarations.
-                                // If the namespace is missing, wrap the element with it.
-                                if (!str_contains($outerXml, 'xmlns')) { // @codeCoverageIgnoreStart
-                                    $outerXml = '<c xmlns="' . $mainNS . '"' . substr($outerXml, 2);
-                                } // @codeCoverageIgnoreEnd
-                                $cellXmlRoot = @simplexml_load_string($outerXml);
-                                if ($cellXmlRoot !== false) {
-                                    $cellXml = $cellXmlRoot->children($mainNS);
-                                }
-                            }
-                        }
-
-                        if (!$this->getReadFilter()->readCell($coordinates[0], (int) $coordinates[1], $docSheet->getTitle())) {
-                            // @codeCoverageIgnoreStart
-                            // Handle shared formulas for filtered cells
-                            if ($cellXml !== null && isset($cellXml->f)) {
-                                $fAttrs = $cellXml->f->attributes();
-                                if (isset($fAttrs['t']) && strtolower((string) $fAttrs['t']) === 'shared') {
-                                    $this->processStreamingFormula($cellXml, $r, $cellDataType, $value, $calculatedValue, false);
-                                }
-                            }
-                            ++$rowIndex;
-
-                            continue;
-                            // @codeCoverageIgnoreEnd
-                        }
-
-                        // Determine if cell contains a formula
-                        $useFormula = false;
-                        if ($cellXml !== null && isset($cellXml->f)) {
-                            $fStr = (string) $cellXml->f;
-                            $fAttrs = $cellXml->f->attributes();
-                            $useFormula = $fStr !== '' || (isset($fAttrs['t']) && strtolower((string) $fAttrs['t']) === 'shared');
-                        }
-
-                        // Get value element text
-                        $vValue = ($cellXml !== null && isset($cellXml->v)) ? (string) $cellXml->v : null;
-
-                        switch ($cellDataType) {
-                            case DataType::TYPE_STRING:
-                                // Shared string
-                                if ($vValue !== null && $vValue !== '') {
-                                    $ssIndex = (int) $vValue;
-                                    $value = $sharedStrings[$ssIndex] ?? '';
-                                    if ($value instanceof RichText) {
-                                        $value = clone $value; // @codeCoverageIgnore
-                                    }
-                                } else {
-                                    $value = ''; // @codeCoverageIgnore
-                                }
-
-                                break;
-
-                            case DataType::TYPE_BOOL:
-                                if (!$useFormula) {
-                                    if ($vValue !== null) {
-                                        $value = ($vValue === '1');
-                                    } else {
-                                        $value = null; // @codeCoverageIgnore
-                                        $cellDataType = DataType::TYPE_NULL; // @codeCoverageIgnore
-                                    }
-                                } else {
-                                    // @codeCoverageIgnoreStart
-                                    $this->processStreamingFormula($cellXml, $r, $cellDataType, $value, $calculatedValue);
-                                    if ($cellXml !== null && isset($cellXml->f)) {
-                                        $this->storeStreamingFormulaAttributes($cellXml->f, $docSheet, $r);
-                                    }
-                                    // @codeCoverageIgnoreEnd
-                                }
-
-                                break;
-
-                            case DataType::TYPE_STRING2:
-                                if ($useFormula) {
-                                    $this->processStreamingFormula($cellXml, $r, $cellDataType, $value, $calculatedValue);
-                                    if ($cellXml !== null && isset($cellXml->f)) {
-                                        $this->storeStreamingFormulaAttributes($cellXml->f, $docSheet, $r);
-                                    }
-                                } else {
-                                    $value = $vValue; // @codeCoverageIgnore
-                                }
-
-                                break;
-
-                            case DataType::TYPE_INLINE:
-                                if ($useFormula) {
-                                    // @codeCoverageIgnoreStart
-                                    $this->processStreamingFormula($cellXml, $r, $cellDataType, $value, $calculatedValue);
-                                    if ($cellXml !== null && isset($cellXml->f)) {
-                                        $this->storeStreamingFormulaAttributes($cellXml->f, $docSheet, $r);
-                                    }
-                                    // @codeCoverageIgnoreEnd
-                                } elseif ($cellXml !== null && isset($cellXml->is)) {
-                                    $value = $this->parseRichText($cellXml->is);
-                                }
-
-                                break;
-
-                            case DataType::TYPE_ERROR:
-                                if (!$useFormula) {
-                                    $value = $vValue;
-                                } else {
-                                    $this->processStreamingFormula($cellXml, $r, $cellDataType, $value, $calculatedValue);
-                                }
-
-                                break;
-
-                            default:
-                                // Numeric or untyped
-                                if (!$useFormula) {
-                                    $value = $vValue;
-                                    if ($value !== null && is_numeric($value)) {
-                                        $value += 0;
-                                        $cellDataType = DataType::TYPE_NUMERIC;
-                                    }
-                                } else {
-                                    $this->processStreamingFormula($cellXml, $r, $cellDataType, $value, $calculatedValue);
-                                    if (is_numeric($calculatedValue)) {
-                                        $calculatedValue += 0;
-                                    }
-                                    if ($cellXml !== null && isset($cellXml->f)) {
-                                        $this->storeStreamingFormulaAttributes($cellXml->f, $docSheet, $r);
-                                    }
-                                }
-
-                                break;
-                        }
-
-                        // Read empty cells or cells that are not empty
-                        if ($this->readEmptyCells || ($value !== null && $value !== '')) {
-                            // Rich text?
-                            if ($value instanceof RichText && $this->readDataOnly) {
-                                $value = $value->getPlainText(); // @codeCoverageIgnore
-                            }
-
-                            $cell = $docSheet->getCell($r);
-                            // Assign value
-                            if ($cellDataType !== '') {
-                                if ($cellDataType === DataType::TYPE_NUMERIC && ($value === '' || $value === null)) {
-                                    $cellDataType = DataType::TYPE_NULL; // @codeCoverageIgnore
-                                }
-                                if ($cellDataType !== DataType::TYPE_NULL) {
-                                    $cell->setValueExplicit($value, $cellDataType);
-                                }
-                            } else {
-                                $cell->setValue($value);
-                            }
-                            if ($calculatedValue !== null) {
-                                $cell->setCalculatedValue($calculatedValue, $originalCellDataTypeNumeric);
-                            }
-
-                            // Style information?
-                            if (!$this->readDataOnly) {
-                                $cAttrS = isset($styles[$styleIndex]) ? $styleIndex : 0;
-                                $cell->setXfIndex($cAttrS);
-                                if ($cellDataType === DataType::TYPE_FORMULA && isset($styles[$cAttrS]) && $styles[$cAttrS]->quotePrefix === true) { //* @phpstan-ignore-line
-                                    $holdSelected = $docSheet->getSelectedCells(); // @codeCoverageIgnore
-                                    $cell->getStyle()->setQuotePrefix(false); // @codeCoverageIgnore
-                                    $docSheet->setSelectedCells($holdSelected); // @codeCoverageIgnore
-                                }
-                            }
-                        }
-
-                        ++$rowIndex;
-                    }
-                }
-
-                ++$cIndex;
-            }
-        }
-
-        $xml->close();
-    }
-
-    /**
-     * Process a formula from a streaming-parsed cell element.
-     */
-    private function processStreamingFormula(
-        ?SimpleXMLElement $cellXml,
-        string $r,
-        string &$cellDataType,
-        mixed &$value,
-        mixed &$calculatedValue,
-        bool $updateSharedCells = true,
-    ): void {
-        if ($cellXml === null || !isset($cellXml->f)) {
-            return; // @codeCoverageIgnore
-        }
-
-        $originalDataType = $cellDataType;
-        $cellDataType = DataType::TYPE_FORMULA;
-        $formula = self::replacePrefixes((string) $cellXml->f);
-        $value = "=$formula";
-
-        // Calculated value from <v>
-        $calculatedValue = isset($cellXml->v) ? (string) $cellXml->v : null;
-        if ($calculatedValue !== null && is_numeric($calculatedValue)) {
-            $calculatedValue += 0;
-        }
-
-        // Handle TYPE_BOOL calculated values - match castToBoolean behavior
-        if ($originalDataType === DataType::TYPE_BOOL && $calculatedValue !== null) {
-            $calculatedValue = (bool) $calculatedValue;
-        }
-
-        // Shared formula?
-        $attr = $cellXml->f->attributes();
-        if (isset($attr['t']) && strtolower((string) $attr['t']) === 'shared') {
-            $instance = (string) $attr['si'];
-
-            if (!isset($this->sharedFormulae[$instance])) {
-                $this->sharedFormulae[$instance] = new SharedFormula($r, $value);
-            } elseif ($updateSharedCells) {
-                $master = Coordinate::indexesFromString($this->sharedFormulae[$instance]->master());
-                $current = Coordinate::indexesFromString($r);
-
-                $difference = [0, 0];
-                $difference[0] = $current[0] - $master[0];
-                $difference[1] = $current[1] - $master[1];
-
-                $value = $this->referenceHelper->updateFormulaReferences(
-                    $this->sharedFormulae[$instance]->formula(),
-                    'A1',
-                    $difference[0],
-                    $difference[1]
-                );
-            }
-        }
-    }
-
-    /**
-     * Store formula attributes from a streaming-parsed formula element.
-     */
-    private function storeStreamingFormulaAttributes(SimpleXMLElement $f, Worksheet $docSheet, string $r): void
-    {
-        $formulaAttributes = [];
-        $attributes = $f->attributes();
-        if (isset($attributes['t'])) {
-            $formulaAttributes['t'] = (string) $attributes['t'];
-        }
-        if (isset($attributes['ref'])) {
-            $formulaAttributes['ref'] = (string) $attributes['ref'];
-        }
-        if (!empty($formulaAttributes)) {
-            $docSheet->getCell($r)->setFormulaAttributes($formulaAttributes);
-        }
-    }
-
-    private static function storeFormulaAttributes(SimpleXMLElement $f, Worksheet $docSheet, string $r): void
+    protected static function storeFormulaAttributes(SimpleXMLElement $f, Worksheet $docSheet, string $r): void
     {
         $formulaAttributes = [];
         $attributes = $f->attributes();
